removed Streaming.Text namespace moved to Streaming directly
